pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sendDetections"
version = "0.1.0"
description = "CSV to Recorded Future Detection API submission utility"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=0.21.0",
    "typing-extensions>=4.1.0",
    "pydantic>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pylint>=2.15.0",
    "mypy>=1.0.0",
    "black>=23.0.0"
]

readme = "README.md"
license = {text = "MIT"}